rename navigation package to api
