+++ app/src/main/java/com/sanaa/novix/di/apiModule.kt
@@ -3,7 +3,7 @@ package com.sanaa.novix.di
 import com.sanaa.api.MediaDetailsApi
 import com.sanaa.api.SearchFeatureApi
 import com.sanaa.presentation.navigation.MediaDetailsApiImpl
-import com.sanaa.presentation.navigation.SearchFeatureApiImpl
+import com.sanaa.presentation.api.SearchFeatureApiImpl
 import org.koin.core.module.dsl.factoryOf
 import org.koin.dsl.bind
 import org.koin.dsl.module
